+++ src/Compute/Compute/Models/PSVirtualMachine.cs
@@ -160,5 +160,8 @@ namespace Microsoft.Azure.Commands.Compute.Models
 
         // Gets or sets the AddProxyAgentExtension
         public bool? AddProxyAgentExtension { get; set; }
+
+        // Gets or sets the ScheduledEventsPolicy
+        public ScheduledEventsPolicy ScheduledEventsPolicy { get; set; }
     }
 }